+++ packages/core/src/index.ts
@@ -1259,17 +1259,27 @@ class Hopr extends EventEmitter {
       await this.strategy.onChannelWillClose(channel)
     }
 
+    // TODO: should remove this blocker when https://github.com/hoprnet/hoprnet/issues/4194 gets addressed
+    if (direction === 'incoming') {
+      log(
+        `Incoming channel: ignoring closing channel ${channel
+          .getId()
+          .toHex()} because current HoprChannels contract does not support closing incoming channels.`
+      )
+      throw new Error('Incoming channel: Closing incoming channels currently is not supported.')
+    }
+
     let txHash: string
     try {
       if (channel.status === ChannelStatus.Open || channel.status == ChannelStatus.WaitingForCommitment) {
         log('initiating closure of channel', channel.getId().toHex())
-        txHash = await connector.initializeClosure(counterpartyPubKey)
+        txHash = await connector.initializeClosure(channel.source, channel.destination)
       } else {
         // verify that we passed the closure waiting period to prevent failing
         // on-chain transactions
 
         if (channel.closureTimePassed()) {
-          txHash = await connector.finalizeClosure(this.pubKey, counterpartyPubKey)
+          txHash = await connector.finalizeClosure(channel.source, channel.destination)
         } else {
           log(
             `ignoring finalizing closure of channel ${channel